Allow users to specify a key for extra fields
Sometimes it is desired to be able to add fields to the HEC payload
without knowing fluentd having to statically know their names in the
<fields> section, such as using annotations to define extra index time
fields. By providing an optional fields_key parameter, users can write
custom filters to add keys to a hash on the record that will then be
added as index time fields on the payload.